Extract OHLCV fetcher and add pagination test
